requirements.txt
google-api-python-client==2.65.0
google-cloud-compute==1.6.1
google-cloud-pubsub==2.13.10
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
